fix: confirmBug 的 pri 強制整數校驗
迴應 code review Minor #2。禪道優先級為整數，但原 schema 僅 z.number()
未限制整數，pri: 2.5 等小數會直接送達禪道。雙層校驗（與 assignedTo 風格一致）：
- API 層（zentao-api.mjs confirmBug）：Number.isInteger 檢查，非整數拋錯且不發請求
- 主伺服器 Zod schema：z.number().int().optional()

補 TDD 測試：pri=2.5 應拋錯且 fetch 不被呼叫。

+++ src/mcp-server.mjs
@@ -443,7 +443,7 @@ server.addTool({
       'codeerror', 'config', 'install', 'security', 'performance',
       'standard', 'automation', 'designdefect', 'others',
     ]).optional().describe('Bug 類型'),
-    pri: z.number().optional().describe('優先級（整數）'),
+    pri: z.number().int().optional().describe('優先級（整數）'),
     comment: z.string().optional().describe('備註說明'),
     mailto: z.array(z.string()).optional().describe('抄送帳號陣列'),
   }),
